requirements.txt
numpy==1.18.1
opencv-contrib-python==4.2.0.32
opencv-python==4.2.0.32
Pillow==7.0.0
tensorflow==2.1.0
plotext==0.1.13
